% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jwatjars-package.R
\docType{package}
\name{jwatjars}
\alias{jwatjars}
\alias{jwatjars-package}
\title{Java '.jar' Files for 'jwatr'}
\description{
By separating out 'JAR' files from 'rJava'-based packages, such packages
can have a lighter filesystem footprint, especially when they are likely to change
at a more rapid pace than the underlying libraries they depend upon. Core library
files from the 'Java Web Archive Toolkit' <https://sbforge.org/display/JWAT/JWAT>
are included and the package version aligns with the underlying 'JAR' release version.
}
\author{
Bob Rudis (bob@rud.is)
}
